backend/requirements.txt
fastapi
uvicorn
redis
google-genai
pydantic
httpx
